Abstract the Book baseline test driver so we can re-use in online docs
