+++ pyomo/common/unittest.py
@@ -17,9 +17,13 @@
 #  ___________________________________________________________________________
 
 import enum
+import glob
 import logging
 import math
+import os
+import operator
 import re
+import subprocess
 import sys
 from io import StringIO
 
@@ -31,7 +35,10 @@ import unittest as _unittest
 import pytest as pytest
 
 from pyomo.common.collections import Mapping, Sequence
+from pyomo.common.dependencies import attempt_import, check_min_version
 from pyomo.common.errors import InvalidValueError
+from pyomo.common.fileutils import import_file
+from pyomo.common.log import LoggingIntercept, pyomo_formatter
 from pyomo.common.tee import capture_output
 
 from unittest import mock
@@ -547,3 +554,405 @@ class TestCase(_unittest.TestCase):
             contextClass = _unittest.case._AssertRaisesContext
         context = contextClass(expected_exception, self, expected_regex)
         return context.handle('assertRaisesRegex', args, kwargs)
+
+
+class BaseLineTestDriver(object):
+    """Generic driver for performing baseline tests in bulk
+
+    This test driver was originally crafted for testing the examples in
+    the Pyomo Book, and has since been generalized to reuse in testing
+    ".. literalinclude:" examples from the Online Docs.
+
+    We expect that consumers of this class will derive from both this
+    class and `pyomo.common.unittest.TestCase`, and then use
+    `parameterized` to declare tests that call either the
+    :py:meth:`python_test_driver` or :py:meth:`shell_test_driver`
+    methods.
+
+    Note that derived classes must declare two class attributes:
+
+    Class Attributes
+    ----------------
+    solver_dependencies: Dict[str, List[str]]
+
+        maps the test name to a list of required solvers.  If any solver
+        is not available, then the test will be skipped.
+
+    package_dependencies: Dict[str, List[str]]
+
+        maps the test name to a list of required modules.  If any module
+        is not available, then the test will be skipped.
+
+    """
+
+    @staticmethod
+    def custom_name_func(test_func, test_num, test_params):
+        func_name = test_func.__name__
+        return "test_%s_%s" % (test_params.args[0], func_name[-2:])
+
+    def __init__(self, test):
+        # Finalize the class, if necessary...
+        if getattr(self.__class__, 'solver_available', None) is None:
+            self.initialize_dependencies()
+        super().__init__(test)
+
+    def initialize_dependencies(self):
+        # Note: aas a rule, pyomo.common is not allowed to import from
+        # the rest of Pyomo.  we permit it here because a) this is not
+        # at module scope, and b) there is really no better / more
+        # logical place in pyomo to put this code.
+        from pyomo.opt import check_available_solvers
+
+        cls = self.__class__
+        #
+        # Initialize the availability data
+        #
+        solvers_used = set(sum(list(cls.solver_dependencies.values()), []))
+        available_solvers = check_available_solvers(*solvers_used)
+        cls.solver_available = {
+            solver_: (solver_ in available_solvers) for solver_ in solvers_used
+        }
+
+        cls.package_available = {}
+        cls.package_modules = {}
+        packages_used = set(sum(list(cls.package_dependencies.values()), []))
+        for package_ in packages_used:
+            pack, pack_avail = attempt_import(package_, defer_check=False)
+            cls.package_available[package_] = pack_avail
+            cls.package_modules[package_] = pack
+
+    @classmethod
+    def _find_tests(cls, test_dirs, pattern):
+        test_tuples = []
+        for testdir in test_dirs:
+            # Find all pattern files in the test directory and any immediate
+            # sub-directories
+            for fname in list(glob.glob(os.path.join(testdir, pattern))) + list(
+                glob.glob(os.path.join(testdir, '*', pattern))
+            ):
+                test_file = os.path.abspath(fname)
+                bname = os.path.basename(test_file)
+                dir_ = os.path.dirname(test_file)
+                name = os.path.splitext(bname)[0]
+                tname = os.path.basename(dir_) + '_' + name
+
+                suffix = None
+                # Look for txt and yml file names matching py file names. Add
+                # a test for any found
+                for suffix_ in ['.txt', '.yml']:
+                    if os.path.exists(os.path.join(dir_, name + suffix_)):
+                        suffix = suffix_
+                        break
+                if suffix is not None:
+                    tname = tname.replace('-', '_')
+                    tname = tname.replace('.', '_')
+
+                    # Create list of tuples with (test_name, test_file, baseline_file)
+                    test_tuples.append(
+                        (tname, test_file, os.path.join(dir_, name + suffix))
+                    )
+
+        # Ensure a deterministic test ordering
+        test_tuples.sort()
+        return test_tuples
+
+    @classmethod
+    def gather_tests(cls, test_dirs):
+        # Find all .sh files in the test directories
+        sh_test_tuples = cls._find_tests(test_dirs, '*.sh')
+
+        # Find all .py files in the test directories
+        py_test_tuples = cls._find_tests(test_dirs, '*.py')
+
+        # If there is both a .py and a .sh, defer to the sh
+        sh_files = set(map(operator.itemgetter(1), sh_test_tuples))
+        py_test_tuples = list(
+            filter(lambda t: t[1][:-3] + '.sh' not in sh_files, py_test_tuples)
+        )
+
+        return py_test_tuples, sh_test_tuples
+
+    def check_skip(self, name):
+        """
+        Return a boolean if the test should be skipped
+        """
+
+        if name in self.solver_dependencies:
+            solvers_ = self.solver_dependencies[name]
+            if not all([self.solver_available[i] for i in solvers_]):
+                # Skip the test because a solver is not available
+                _missing = []
+                for i in solvers_:
+                    if not self.solver_available[i]:
+                        _missing.append(i)
+                return "Solver%s %s %s not available" % (
+                    's' if len(_missing) > 1 else '',
+                    ", ".join(_missing),
+                    'are' if len(_missing) > 1 else 'is',
+                )
+
+        if name in self.package_dependencies:
+            packages_ = self.package_dependencies[name]
+            if not all([self.package_available[i] for i in packages_]):
+                # Skip the test because a package is not available
+                _missing = []
+                for i in packages_:
+                    if not self.package_available[i]:
+                        _missing.append(i)
+                return "Package%s %s %s not available" % (
+                    's' if len(_missing) > 1 else '',
+                    ", ".join(_missing),
+                    'are' if len(_missing) > 1 else 'is',
+                )
+
+            # This is a hack, xlrd dropped support for .xlsx files in 2.0.1 which
+            # causes problems with older versions of Pandas<=1.1.5 so skipping
+            # tests requiring both these packages when incompatible versions are found
+            if (
+                'pandas' in self.package_dependencies[name]
+                and 'xlrd' in self.package_dependencies[name]
+            ):
+                if check_min_version(
+                    self.package_modules['xlrd'], '2.0.1'
+                ) and not check_min_version(self.package_modules['pandas'], '1.1.6'):
+                    return "Incompatible versions of xlrd and pandas"
+
+        return False
+
+    def filter_fcn(self, line):
+        """
+        Ignore certain text when comparing output with baseline
+        """
+        for field in (
+            '[',
+            'password:',
+            'http:',
+            'Job ',
+            'Importing module',
+            'Function',
+            'File',
+            'Matplotlib',
+            '-------',
+            '=======',
+            '    ^',
+        ):
+            if line.startswith(field):
+                return True
+        for field in (
+            'Total CPU',
+            'Ipopt',
+            'license',
+            'Status: optimal',
+            'Status: feasible',
+            'time:',
+            'Time:',
+            'with format cpxlp',
+            'usermodel = <module',
+            'execution time=',
+            'Solver results file:',
+            'TokenServer',
+            'function calls',
+            'List reduced',
+            '.py:',
+            '{built-in method',
+            '{method',
+            '{pyomo.core.expr.numvalue.as_numeric}',
+        ):
+            if field in line:
+                return True
+        return False
+
+    def filter_file_contents(self, lines):
+        filtered = []
+        deprecated = None
+        for line in lines:
+            # Ignore all deprecation warnings
+            if line.startswith('WARNING: DEPRECATED:'):
+                deprecated = ''
+            if deprecated is not None:
+                deprecated += line
+                if re.search(r'\(called\s+from[^)]+\)', deprecated):
+                    deprecated = None
+                continue
+
+            if not line or self.filter_fcn(line):
+                continue
+
+            # Strip off beginning of lines giving time in seconds
+            # Needed for the performance chapter tests
+            if "seconds" in line:
+                s = line.find("seconds") + 7
+                line = line[s:]
+
+            items = line.strip().split()
+            for i in items:
+                if not i:
+                    continue
+                if i.startswith('/') or i.startswith(":\\", 1):
+                    continue
+
+                # A few substitutions to get tests passing on pypy3
+                if ".inf" in i:
+                    i = i.replace(".inf", "inf")
+                if "null" in i:
+                    i = i.replace("null", "None")
+
+                try:
+                    filtered.append(float(i))
+                except:
+                    filtered.append(i)
+
+        return filtered
+
+    def compare_baselines(self, test_output, baseline, abstol=1e-6, reltol=None):
+        # Filter files independently and then compare filtered contents
+        out_filtered = self.filter_file_contents(test_output.strip().split('\n'))
+        base_filtered = self.filter_file_contents(baseline.strip().split('\n'))
+
+        if len(out_filtered) != len(base_filtered):
+            # it is likely that a solver returned a (slightly) nonzero
+            # value for a variable that is normally 0.  Try to look for
+            # sequences like "['varname:', 'Value:', 1e-9]" that appear
+            # in one result but not the other and remove them.
+            i = 0
+            while i < min(len(out_filtered), len(base_filtered)):
+                try:
+                    self.assertStructuredAlmostEqual(
+                        out_filtered[i],
+                        base_filtered[i],
+                        abstol=abstol,
+                        reltol=reltol,
+                        allow_second_superset=False,
+                    )
+                    i += 1
+                    continue
+                except self.failureException:
+                    pass
+
+                try:
+                    index_of_out_i_in_base = base_filtered.index(out_filtered[i], i)
+                except ValueError:
+                    index_of_out_i_in_base = float('inf')
+                try:
+                    index_of_base_i_in_out = out_filtered.index(base_filtered[i], i)
+                except ValueError:
+                    index_of_base_i_in_out = float('inf')
+                if index_of_out_i_in_base < index_of_base_i_in_out:
+                    extra = base_filtered
+                    n = index_of_out_i_in_base
+                else:
+                    extra = out_filtered
+                    n = index_of_base_i_in_out
+                if n == float('inf'):
+                    n = None
+                extra_terms = extra[i:n]
+                try:
+                    assert len(extra_terms) % 3 == 0
+                    assert all(str(_)[-1] == ":" for _ in extra_terms[0::3])
+                    assert all(str(_) == "Value:" for _ in extra_terms[1::3])
+                    assert all(abs(_) < abstol for _ in extra_terms[2::3])
+                except:
+                    # This does not match the pattern we are looking
+                    # for: quit processing, and let the next
+                    # assertStructuredAlmostEqual raise the appropriate
+                    # failureException
+                    break
+                extra[i:n] = []
+
+        try:
+            self.assertStructuredAlmostEqual(
+                out_filtered,
+                base_filtered,
+                abstol=abstol,
+                reltol=reltol,
+                allow_second_superset=False,
+            )
+        except self.failureException:
+            # Print helpful information when file comparison fails
+            print('---------------------------------')
+            print('BASELINE FILE')
+            print('---------------------------------')
+            print(baseline)
+            print('=================================')
+            print('---------------------------------')
+            print('TEST OUTPUT FILE')
+            print('---------------------------------')
+            print(test_output)
+            raise
+        return True
+
+    def python_test_driver(self, tname, test_file, base_file):
+        bname = os.path.basename(test_file)
+        _dir = os.path.dirname(test_file)
+
+        skip_msg = self.check_skip('test_' + tname)
+        if skip_msg:
+            raise _unittest.SkipTest(skip_msg)
+
+        with open(base_file, 'r') as FILE:
+            baseline = FILE.read()
+
+        cwd = os.getcwd()
+        try:
+            os.chdir(_dir)
+            # This is roughly equivalent to:
+            #    subprocess.run([sys.executable, bname],
+            #                   stdout=f, stderr=f, cwd=_dir)
+            with capture_output(None, True) as OUT:
+                # Note: we want LoggingIntercept to log to the
+                # *current* stdout (which is the TeeStream from
+                # capture_output).  This ensures that log messages and
+                # normal output appear in the correct order.
+                with LoggingIntercept(sys.stdout, formatter=pyomo_formatter):
+                    import_file(bname, infer_package=False, module_name='__main__')
+        finally:
+            os.chdir(cwd)
+
+
+        try:
+            self.compare_baselines(OUT.getvalue(), baseline)
+        except:
+            if os.environ.get('PYOMO_TEST_UPDATE_BASELINES', None):
+                with open(base_file, 'w') as FILE:
+                    FILE.write(OUT.getvalue())
+            raise
+
+    def shell_test_driver(self, tname, test_file, base_file):
+        bname = os.path.basename(test_file)
+        _dir = os.path.dirname(test_file)
+
+        skip_msg = self.check_skip('test_' + tname)
+        if skip_msg:
+            raise _unittest.SkipTest(skip_msg)
+
+        # Skip all shell tests on Windows.
+        if os.name == 'nt':
+            raise _unittest.SkipTest("Shell tests are not runnable on Windows")
+
+        with open(base_file, 'r') as FILE:
+            baseline = FILE.read()
+
+        cwd = os.getcwd()
+        try:
+            os.chdir(_dir)
+            _env = os.environ.copy()
+            _env['PATH'] = os.pathsep.join(
+                [os.path.dirname(sys.executable), _env['PATH']]
+            )
+            rc = subprocess.run(
+                ['bash', bname],
+                stdout=subprocess.PIPE,
+                stderr=subprocess.STDOUT,
+                cwd=_dir,
+                env=_env,
+            )
+        finally:
+            os.chdir(cwd)
+
+        try:
+            self.compare_baselines(rc.stdout.decode(), baseline)
+        except:
+            if os.environ.get('PYOMO_TEST_UPDATE_BASELINES', None):
+                with open(base_file, 'w') as FILE:
+                    FILE.write(rc.stdout.decode())
+            raise
